+++ src/app/models/product-category.enum.ts
@@ -1,5 +1,0 @@
-export enum ProductCategory {
-  First,
-  Second,
-  Thrid,
-}